docs(look-and-feel): skeleton wrapper mode
